Adds support for `~/.capfile`, closes #445
Any tasks specified in `~/.capfile` will be loaded when including
`capistrano/setup`.  Given the default nature of these tasks, it may be
that this load order is not ideal, but for the simple use case of having
a number of default tasks available to all applications I think this is
sufficient.  It would however be preferable to encourage the packaging
of shared tasks in gem format.