Alter the default template to make it clear how to specify the deploy username.
This is preferred over the old style of setting a single deploy username, as
different servers may require different logins, for example database servers
typically are easier to work with if logging in with a `postgres` or similar
user.

This also allows people to simply specify `myserver`, and rely on the host SSH
configuration to make that work.