Revert ":panda_face: 857: Load tasks from the deploy.rb"
This reverts commit a5b76f782a11b191aff87e52734202c949dbbb3b.

This commit which looked to solve #867 / #857 has caused regressions in
other areas - I'm reverting this commit because the solution to the original
issue of tasks declared in `config/deploy.rb` not appearing in the list of
tasks returned by `cap -T` is to move those tasks into
`lib/capistrano/tasks` and load them from the Capfile (as per the
generated template). Defining tasks in a configuration file may work as
far as running the tasks goes, but it's not the intended use for that
file.

This commit closes #887, and I suspect #881 as well.