Allow lambdas with arity > 0 as config options
Previously it was possible to `set` a lambda with one or more
parameters as a configuration option, but at runtime an error was
thrown when `fetch`ing it. This was caused by the fact that every
callable was directly invoke in `fetch`.

This PR allows for lambdas with an arity > 0 to be retrieved
through `fetch`. All values that respond to call and NOT to
`arity` are still invoked directly by `fetch`.