Remove unnecessary entries from default backtrace
When the --backtrace (or --trace) command line option is not supplied
Rake lowers the noise level in exception backtraces by building
a regular expression containing all the system library paths and
using it to exclude backtrace entries that match.

This doesn't always go far enough, particularly in RVM environments when
many gem paths are added. This commit reverses that approach and only
include backtrace entries that fall within the Capfile and list of tasks
imported thereafter.

If the full unexpurgated backtrace is required then the --backtrace
and --trace options supply it as before.