Allow the tar command name to be overridden
When using the git integration, it generates a tarball and then untars it to generate the release folder. This has been hardcoded to use `tar` up until now, but some OS' provide a tar that isn't gnu compatible (eg, SmartOS). Instead they have `gtar` available.

This is now picked up from the SSHKit's command map, so it can be overridden as required:

    SSHKit.config.command_map[:tar] = "/usr/bin/env gtar"