Create new SCM plugin system; deprecate `set :scm`
This commit implements the migration path from the old way of loading the SCM
implementation, which is to say, from this:

set :scm, :git

To this:

require "capistrano/scm/git"
install_plugin Capistrano::SCM::Git

I do this in such a way that all existing Capistrano projects should continue to
work unmodified. Internally the new SCM plugin will be used automatically and a
deprecation notice will be printed. For example, if you are using
set :scm, :git, you will see:

[Deprecation Notice] `set :scm, :git` is deprecated.
To ensure your project is compatible with future versions of Capistrano,
remove the :scm setting and instead add these lines to your Capfile:

    require "capistrano/scm/git"
    install_plugin Capistrano::SCM::Git

If you are using a third-party SCM, like :copy, then you would be given a link
to documentation on the topic:

[Deprecation Notice] `set :scm, :copy` is deprecated.
To ensure this custom SCM will work with future versions of Capistrano,
please upgrade it to a version that uses the new SCM plugin mechanism
documented here:

http://capistranorb.com/documentation/advanced-features/custom-scm

All of this logic is pretty much self-contained in a single class: SCMResolver,
which can go away once Capistrano 4.0 is released and we can drop backwards
compatibility.