Mute RuboCop method_missing check
RuboCop wants us to call `super` inside `method_missing`, which is
entirely reasonable 99% of the time. But in this case our implementation
is swallowing *all* possible method names, so there is no opportunity
to delegate to `super`.

This may be a design flaw in Capistrano, but I think that ship has
sailed.