Ensure validation works for callable values
The previous solution worked for validating values set using the block
syntax, like this: `set(:application) { "value" }`. However it failed
for the much more common scenario of `set :application, -> { "value" }`.

To fix this, we have to check whether the value passed to set is
callable (i.e. responds to `call` with no arguments). If so, we apply
the deferred validation logic in this scenario as well.

The additional wrinkle is the case where the value is a `Question`
object. A Question responds to `call`, but it is not a normal Ruby Proc.
We can't simply wrap it in another Proc to add validation because then
`Configuration#is_question?` would erroneously return false.

To account for this, I use a trivial `ValidatedQuestion` class that
simply delegates to the Proc that does the validation. This class is a
subclass of `Question`, so the `Configuration#is_question?` test still
works.

Finally, in an effort to keep backtraces as clean as possible, all of
this validation wrapping stuff is only applied if a validation rule is
actually registered. If no validation rules are defined for a given key,
there is no benefit to this meta-programming.