Exclude docs from gem package
Now that the entire Capistrano docs website is contained in the docs
directory, by default all those files will be included in the .gem file
that is created by `rake build`.

This means that the capistrano gem has increased in size from 68 KB to
802 KB. More than 10x!

This commit updates the gemspec to include the docs directory when
building the gem, to avoid this bloat.