Fix test suite for mocha 1.2.0
This comes down to a behaviour change in mocha not playing so nicely with how we've been generating empty mock objects so far.

Mocha 1.2.0 now pays attention to the visibility of methods you stub (expect(), etc) and disallows you calling them in a different scope. ie, if you stub a private method it can't be called from outside the instance. Previously doing that worked fine and you had no indication you were calling a private method. This change totally makes sense!

However, we're providing "empty" contexts in our specs by using `Class.new.new` which is creating an anonymous instance of an anonymous class. Note that the anonymous class contains Kernel in the ancestors.

We then stub out `test` on these anonymous instances, which we don't expect to respond to `test` without the stubbing, and the stubbed method then gets called. However, remember back to where we had `Kernel` methods available in our anonymous instance? Turns out there's `Kernel#test` and it's marked private! Except that our method on the real context instances is public, so we want that to work.

Easiest fix I can see is to use Mocha's `mock` method to create a mock object for us to stub out our methods on, rather than trying to be ruby-ish and use an anonymous instance of an anonymous class. (I attempted using BasicObject first, but we'd have to include Mocha into that ourselves, and it leans on some methods provided by Kernel.)