Fix hg repo_tree implementation
`hg archive` does not write to standard out, so Capistrano's `git`
implementation of the `:repo_tree` option does not work for `hg`. As a
workaround, generate a temporary tar file, extract it, and then remove
it. This is not as efficient or elegant as doing it via stdout and pipes
in a single line, but it should work.