Maintain release notes using GitHub Actions (#2034)
This commit sets up a GitHub Actions workflow that uses Release Drafter
to automatically maintain release notes on every push.

In practice this means that contributors no longer have to manually
update the CHANGELOG, which is something that is easy to forget and
often introduces tedious merge conflicts.

Instead, Release Drafter automatically adds the title of the PR to the
GitHub release notes to a draft release, crediting the author of the PR,
and linking to the PR number. Release Drafter furthermore organizes the
release notes into sections according to the labels assigned to the PRs:

- ⚠️ Breaking
- 🐛 Bug Fix
- 📚 Docs
- ✨ Feature
- 🏠 Housekeeping

This also simplifies the release process for maintainers: instead of
manually updating the CHANGELOG, all you have to do is press "publish"
on the release draft on GitHub.

This brings capistrano into parity with the configuration of sshkit.